pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["g0dm0d <ivan.berlin06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
python-multipart = "^0.0.9"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
